+++ src/mpmToolbox/gui/mpmEditingTools/editDialogs/ornamentDef/DynamicsGradientPanel.java
@@ -31,7 +31,7 @@ public class DynamicsGradientPanel extends WebPanel {
     private final WebSpinner transitionToSpinner;
     private final WebTextField id = new WebTextField();
     private final WebButton generateId = new WebButton("Generate");
-    private final WebLabel idLabel= new WebLabel("ID:");
+    private final WebLabel idLabel= new WebLabel("ID");
 
     /**
      * constructor
@@ -41,16 +41,19 @@ public class DynamicsGradientPanel extends WebPanel {
         GridBagLayout layout = new GridBagLayout();
         this.setLayout(layout);
 
-        this.setPadding(Settings.paddingInDialogs);
-        this.setBorder(BorderFactory.createCompoundBorder(new LineBorder(Color.BLACK), new EmptyBorder(Settings.paddingInDialogs, Settings.paddingInDialogs, Settings.paddingInDialogs, Settings.paddingInDialogs)));
+//        this.setPadding(Settings.paddingInDialogs);
+        this.setBorder(BorderFactory.createCompoundBorder(new LineBorder(this.getBackground(), Settings.paddingInDialogs/2), new EmptyBorder(Settings.paddingInDialogs, Settings.paddingInDialogs, Settings.paddingInDialogs, Settings.paddingInDialogs)));
+        this.setBackground(Tools.brighter(this.getBackground(), 0.07));
 
         // visualizer
         DynamicsGradientVisualizer dynamicsGradientVisualizer = new DynamicsGradientVisualizer();
-        Tools.addComponentToGridBagLayout(this, layout, dynamicsGradientVisualizer, 3, 0, 1, 1, 20.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        dynamicsGradientVisualizer.setToolTip("accentuation gradient of the ornament");
+        Tools.addComponentToGridBagLayout(this, layout, dynamicsGradientVisualizer, 1, 0, 1, 1, 4.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
 
         // transitionFrom slider
         this.transitionFromSlider = new WebSlider(WebSlider.VERTICAL, -100000, 100000, 0);
         this.transitionFromSlider.setPadding(0, 0, 0, 0);
+        this.transitionFromSlider.setToolTip("accentuation level at the beginning of the ornament");
         this.transitionFromSlider.setMinimumHeight(Settings.getDefaultFontSize() * 3);
         this.transitionFromSlider.setPreferredHeight(this.transitionFromSlider.getMinimumHeight());
         this.transitionFromSlider.setMaximumHeight(this.transitionFromSlider.getMinimumHeight());
@@ -83,12 +86,12 @@ public class DynamicsGradientPanel extends WebPanel {
             public void mouseExited(MouseEvent e) {
             }
         });
-        Tools.addComponentToGridBagLayout(this, layout, this.transitionFromSlider, 1, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        Tools.addComponentToGridBagLayout(this, layout, this.transitionFromSlider, 0, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
 
-        this.transitionFromSpinner = new WebSpinner(new SpinnerNumberModel(0, -1.0, 1.0, 0.1));
+        this.transitionFromSpinner = new WebSpinner(new SpinnerNumberModel(0.0, -1.0, 1.0, 0.1));
         this.transitionFromSpinner.setMinimumWidth(this.SPINNER_WIDTH);
         this.transitionFromSpinner.setMaximumWidth(this.SPINNER_WIDTH);
-        Tools.addComponentToGridBagLayout(this, layout, this.transitionFromSpinner, 2, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        Tools.addComponentToGridBagLayout(this, layout, this.transitionFromSpinner, 0, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
 
         this.transitionFromSlider.addChangeListener(changeEvent -> {
             double value = Tools.round(((double) this.transitionFromSlider.getValue()) / 100000.0, 10);
@@ -104,6 +107,7 @@ public class DynamicsGradientPanel extends WebPanel {
         // transitionTo slider
         this.transitionToSlider = new WebSlider(WebSlider.VERTICAL, -100000, 100000, 0);
         this.transitionToSlider.setPadding(0, 0, 0, 0);
+        this.transitionToSlider.setToolTip("accentuation level at the end of the ornament");
         this.transitionToSlider.setMinimumHeight(Settings.getDefaultFontSize() * 3);
         this.transitionToSlider.setPreferredHeight(this.transitionToSlider.getMinimumHeight());
         this.transitionToSlider.setMaximumHeight(this.transitionToSlider.getMinimumHeight());
@@ -139,12 +143,12 @@ public class DynamicsGradientPanel extends WebPanel {
         this.transitionToSlider.addChangeListener(changeEvent -> {
             dynamicsGradientVisualizer.setTransitionTo(Tools.round(((double) this.transitionToSlider.getValue()) / 100000.0, 10));
         });
-        Tools.addComponentToGridBagLayout(this, layout, this.transitionToSlider, 5, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        Tools.addComponentToGridBagLayout(this, layout, this.transitionToSlider, 2, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
 
-        this.transitionToSpinner = new WebSpinner(new SpinnerNumberModel(0, -1.0, 1.0, 0.1));
+        this.transitionToSpinner = new WebSpinner(new SpinnerNumberModel(0.0, -1.0, 1.0, 0.1));
         this.transitionToSpinner.setMinimumWidth(this.SPINNER_WIDTH);
         this.transitionToSpinner.setMaximumWidth(this.SPINNER_WIDTH);
-        Tools.addComponentToGridBagLayout(this, layout, this.transitionToSpinner, 4, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        Tools.addComponentToGridBagLayout(this, layout, this.transitionToSpinner, 2, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
 
         this.transitionToSlider.addChangeListener(changeEvent -> {
             double value = Tools.round(((double) this.transitionToSlider.getValue()) / 100000.0, 10);
@@ -158,32 +162,15 @@ public class DynamicsGradientPanel extends WebPanel {
         });
 
         // id input
-        Tools.addComponentToGridBagLayout(this, layout, this.makeIdPanel(), 0, 1, 6, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-    }
-
-    /**
-     * generate an ID panel for the sub-elements of the ornamentDef
-     * @return
-     */
-    private WebPanel makeIdPanel() {
-        GridBagLayout layout = new GridBagLayout();
-        WebPanel idPanel = new WebPanel(layout);
-        idPanel.setPadding(Settings.paddingInDialogs, 0, 0, 0);
-
-        this.idLabel.setHorizontalAlignment(WebTextField.LEFT);
-//        idLabel.setPadding(Settings.paddingInDialogs, 0, Settings.paddingInDialogs, 0);
-//        idLabel.setBackground(transparent);
-        Tools.addComponentToGridBagLayout(idPanel, layout, this.idLabel, 0, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-
+        this.idLabel.setHorizontalAlignment(WebLabel.RIGHT);
+        this.idLabel.setPadding(0, 0, 0, Settings.paddingInDialogs);
+        Tools.addComponentToGridBagLayout(this, layout, this.idLabel, 0, 2, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
         this.id.setMaximumWidth(this.getFontMetrics(this.id.getFont()).stringWidth("wwwwwwwwwww"));
         this.id.setMinimumWidth(this.getFontMetrics(this.id.getFont()).stringWidth("wwwwwwwwwww"));
-        Tools.addComponentToGridBagLayout(idPanel, layout, this.id, 1, 0, 1, 1, 8.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-
+        Tools.addComponentToGridBagLayout(this, layout, this.id, 1, 2, 1, 1, 4.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
         this.generateId.setHorizontalAlignment(WebButton.CENTER);
         this.generateId.addActionListener(actionEvent -> this.id.setText(UUID.randomUUID().toString()));
-        Tools.addComponentToGridBagLayout(idPanel, layout, this.generateId, 2, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-
-        return idPanel;
+        Tools.addComponentToGridBagLayout(this, layout, this.generateId, 2, 2, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
     }
 
     /**