+++ history.md
@@ -5,6 +5,7 @@
 - When initializing an `Alignment` (in the audio tab) not only will the last note be fixed but now also the first note. So wherever we start shifting notes, it will act more comprehensible now.
 - Added handling/conversions of dates if the pulses per quarter value of the MSM and the performance differ. This addresses issue [14](https://github.com/axelberndt/MPM-Toolbox/issues/14). There are, however, still some "nitpicks" left with the TempoMap panel.
 - Added default initialization to the Timing Basis field in the Performance Editor dialog, so it matches the `pulsesPerQuarter` value of the MSM.
+- Added `ornamentDef` editor dialog and functionality.
 
 #### v0.1.20
 - The MIDI player of the `mpmToolbox.gui.ProjectPane` has been moved to class `mpmToolbox.gui.MpmToolbox`, so it does not have to be reinitialized when loading a new project. The same has been done with the `mpmToolbox.gui.syncPlayer.SyncPlayer`'s MIDI player. When loading a soundfont, only the latter will change. The MSM tree's MIDI player does play only with piano sounds anyway.